+++ scripts/module.js
@@ -12,12 +12,13 @@ import { shakeScreen } from "./helpers/animation/shakeScreen.js";
 import { generateRollScroll } from "./helpers/animation/generateRollScroll.js";
 import { generateDamageScroll } from "./helpers/animation/generateDamageScroll.js";
 import { getDamageList } from "./helpers/rollTerms.js";
-import { injectConfig } from "./helpers/injectConfig.js";
 import { createFinishingMoveAnimation } from "./helpers/animation/finishingMove.js";
 import { createCritAnimation } from "./helpers/animation/crit/critAnimation.js";
 import { sendUpdateMessage } from "./helpers/tours/updateMessage.js";
 import { createAPI } from "./helpers/api.js";
 import { createBasicActionAnimation } from "./helpers/animation/basicActionAnimation.js";
+import { setupTokenMenu } from "./helpers/UI/tokenUI.js";
+import { shakeOnAttack } from "./helpers/animation/shakeScreenOnAttack.js";
 
 // HOOKS STUFF
 Hooks.on("init", () => {
@@ -70,8 +71,13 @@ Hooks.on("ready", () => {
             // RPG Numbers on Check Roll
             checkRollNumbers(dat, msg);
 
-            // Rotate on Attack Roll
-            if (isRotateOnAttack(dat)) rotateOnAttack(msg);
+            //Attack Roll Stuff
+            if (dat.isAttackRoll) {
+                // Rotate on Attack Roll
+                if (isRotateOnAttack()) rotateOnAttack(msg);
+                if (isShakeOnAttack(msg.token.actor))
+                    handleDiceSoNice(shakeOnAttack, [msg.token, msg.flags.pf2e.context.outcome], msg);
+            }
 
             //On Damage Application
             onDamageApplication(dat, msg);
@@ -103,65 +109,6 @@ Hooks.on("ready", () => {
     console.log("PF2e RPG Numbers is ready");
 });
 
-function setupTokenMenu() {
-    injectConfig.quickInject([{ documentName: "Token" }], {
-        moduleId: MODULE_ID,
-        tab: {
-            name: MODULE_ID,
-            label: localize("token-options.tab-label"),
-            icon: "fas fa-dragon",
-        },
-        rotationOffset: {
-            type: "number",
-            label: localize("token-options.rotation-offset.name"),
-            notes: localize("token-options.rotation-offset.hint"),
-            default: 0,
-        },
-        fireEmblemImg: {
-            type: "filepicker",
-            label: localize("token-options.fire-emblem-img.name"),
-            notes: localize("token-options.fire-emblem-img.hint"),
-            default: "",
-        },
-        personaImg: {
-            type: "filepicker",
-            label: localize("token-options.persona-img.name"),
-            notes: localize("token-options.persona-img.hint"),
-            default: "",
-        },
-        critOffsetX: {
-            type: "number",
-            label: localize("token-options.crit.offset-x.name"),
-            notes: localize("token-options.crit.offset-x.hint"),
-            default: 0,
-        },
-        critOffsetY: {
-            type: "number",
-            label: localize("token-options.crit.offset-y.name"),
-            notes: localize("token-options.crit.offset-y.hint"),
-            default: 0,
-        },
-        critScale: {
-            type: "number",
-            label: localize("token-options.crit.scale.name"),
-            notes: localize("token-options.crit.scale.hint"),
-            default: 100,
-        },
-        critRotation: {
-            type: "number",
-            label: localize("token-options.crit.rotation.name"),
-            notes: localize("token-options.crit.rotation.hint"),
-            default: 0,
-        },
-        critSFX: {
-            type: "filepicker.audio",
-            label: localize("token-options.crit.sfx.name"),
-            notes: localize("token-options.crit.sfx.hint"),
-            default: "",
-        },
-    });
-}
-
 //createCritAnimation({ type: "custom", whisper: [game.user.id], token: token ?? game.user.character });
 
 function getData(msg) {
@@ -179,7 +126,7 @@ function getData(msg) {
 }
 
 function onDamageApplication(dat, msg) {
-    if (dat.isApplyDamage && doSomethingOnDamageApply) {
+    if (dat.isApplyDamage && doSomethingOnDamageApply()) {
         const dmg = dat.appliedDamage.updates.find((u) => u.path === "system.attributes.hp.value")?.value;
         if (dmg) {
             activateShakeToken(dat, dmg);
@@ -203,8 +150,22 @@ function activateShakeToken(dat, dmg) {
         shakeOnDamageToken(dat.appliedDamage?.uuid, dmg);
 }
 
-function isRotateOnAttack(dat) {
-    return dat.isAttackRoll && getSetting("rotate-on-attack");
+function isShakeOnAttack(actor) {
+    if (!getSetting("shake-on-attack.enabled")) return false;
+    const isPlayerOwned = actor.hasPlayerOwner;
+    switch (getSetting("shake-on-attack.type")) {
+        case "gm":
+            return !isPlayerOwned;
+        case "players":
+            return isPlayerOwned;
+        case "both":
+        default:
+            return true;
+    }
+}
+
+function isRotateOnAttack() {
+    return getSetting("rotate-on-attack");
 }
 function rotateOnAttack(msg) {
     turnTokenOnAttack(msg?.token?.object, msg?.target?.token?.object);